Read in SAS datasets in .sas7bdat format into Spark by using the spark-sas7bdat Spark package.

Read in SAS datasets in .sas7bdat format into Spark by using the spark-sas7bdat Spark package.

spark_read_sas(sc, path, table)

Arguments

sc

Connection to Spark local instance or remote cluster. See the example

path

full path to the SAS file either on HDFS (hdfs://), S3 (s3n://), as well as the local file system (file://). Mark that files on the local file system need to be specified using the full path.

table

character string with the name of the Spark table where the SAS dataset will be put into

Value

an object of class tbl_spark, which is a reference to a Spark DataFrame based on which dplyr functions can be executed. See https://github.com/rstudio/sparklyr

References

https://spark-packages.org/package/saurfang/spark-sas7bdat, https://github.com/saurfang/spark-sas7bdat, https://github.com/rstudio/sparklyr

See also

spark_connect, sdf_register

Examples

## Not run: ------------------------------------
# myfile <- system.file("extdata", "iris.sas7bdat", package = "spark.sas7bdat")
# myfile
# 
# library(sparklyr)
# library(spark.sas7bdat)
# sc <- spark_connect(master = "local")
# x <- spark_read_sas(sc, path = myfile, table = "sas_example")
# x
## ---------------------------------------------